Create surface_free_eng.py
This script first prompts the user to enter the molar mass and surface area of the material using the input function. Then, it calculates the molar volume using the molar mass and Avogadro's number. After that, it uses the molar volume and surface area to calculate the surface free energy using the formula:
(2 * (R * T * ln(surface_area / molar_volume)) / (surface_area * 10^-4))
Finally, it prints the result of surface free energy in J/cm^2.
This script assumes that the temperature is at 298K, but you can change it to any other value you want.